Fix size estimate in bidding and auction serializer
We were comparing the value of a uint_64t to 0xFFFFFFFF + 1 + 4, which the C++ compiler decides is equal to 4, instead of 0x100000004. Add type annotation to the constant to specify that it really is a 64 bit number.

Bug: 374327705
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5948304
Commit-Queue: mmenke <mmenke@chromium.org>
Auto-Submit: Russ Hamilton <behamilton@google.com>
Cr-Commit-Position: refs/heads/main@{#1371613}

+++ content/browser/interest_group/bidding_and_auction_serializer.cc
@@ -134,7 +134,7 @@ size_t MaxLengthOfTaggedData(uint64_t length) {
   } else if (length <= 0xFFFF + 1 + 2) {
     // 1 byte tag, 2 bytes length.
     lol_x = 2;
-  } else if (length <= 0xFFFFFFFF + 1 + 4) {
+  } else if (length <= static_cast<uint64_t>(0xFFFFFFFF) + 1 + 4) {
     // 1 byte tag, 4 bytes length.
     lol_x = 4;
   } else {
